Make start timeout work properly.
self.alive is an RPC call, which can take a long time (e.g. 45s) if there are problems with communicating to server.
Previously the loop was being executed exactly timeout * 10 times befor timing out, resulting in 450s timeout instead of the requested value.
Now it is executed only until current timestamp is larger or equal the pre-calculated expiration time.